Organized the Project according to standard conventions and implemented email verification feature
