+++ requirements.txt
@@ -1,3 +1,5 @@
 Flask
 flask-SQLAlchemy
+flask-mail
 mysqlclient
+itsdangerous
